Refactor `generate_output` in anticipation of APProcedurePatch
* Move most ROM-related code to `Rom.py`
* Replace `openFile` and related functions with native `pkgutil.get_data`
